user_service/requirements.txt
Flask
pymongo
PyJWT 
Flask-CORS